Text Display on Gaze
Text is displayed accordingly  when hit by the cursor. The position and rotation of the text follows the cursor. Issues:  The rotation needs to be fixed as currently it's inverted. The transform updating of the text is  linked to DefaultCursor transfrom object, but there must be a better way to achieve that. through MRTK API.