Text cleanup
Implements multiple things. First it uses TextMeshPro to generate text with a lot more control like leading and kerning of characters. The class TextDisplay no longer exists, it is now replaced by GraphInteraction. This class highlights the bars with a selected material when focus with the cursor, while also displaying the text accordingly, and updating the position according to the user's gaze. The issue of  the cursor changed when using the control still exists but it might not be an issue because having two cursors is not possible and gazing without a cursor is very confusing, thus I believe it is npt a priority.